feat: Extracted class for xml from IdeaUtils
